server/src/requirements.txt
certifi==2017.11.5
Django==2.0.1
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2017.3
wincertstore==0.2
